Initial version of the drapak-reddit app rewritten
Initial version of the app. The app builds and runs but needs improvements in the logic to use Hibernate more efficiently. Sql scripts have been changed compared to the drapak-reddit and the all the properties files have been moved to the resources.  Repositories and Controllers work. Logic for the db must be changed, caching and tests to be added.